+++ lib/bunny_publisher/base.rb
@@ -6,6 +6,14 @@ module BunnyPublisher
   class Base
     include Callbacks
 
+    # A list of errors that can be fixed by a connection recovery
+    RETRIABLE_ERRORS = [
+      Bunny::ConnectionClosedError,
+      Bunny::NetworkFailure,
+      Bunny::ConnectionLevelException,
+      Timeout::Error # can be raised by Bunny::Channel#with_continuation_timeout
+    ].freeze
+
     attr_reader :connection, :channel, :exchange
 
     def initialize(publish_connection: nil, connection: nil, exchange: nil, exchange_options: {}, **options)
@@ -27,10 +35,11 @@ module BunnyPublisher
         @message = message
         @message_options = message_options
 
-        ensure_connection!
-
         run_callbacks(:publish) do
-          exchange.publish(message, message_options.dup) # Bunny modifies message options
+          with_errors_handling do
+            ensure_connection!
+            exchange.publish(message, message_options.dup) # Bunny modifies message options
+          end
         end
       ensure
         @message = @message_options = nil
@@ -47,27 +56,64 @@ module BunnyPublisher
 
     attr_reader :message, :message_options
 
+    delegate :logger, to: :connection
+
     def ensure_connection!
       @connection ||= build_connection
 
-      connection.start if connection.status == :not_connected # Lazy connection initialization.
+      connection.start if should_start_connection?
 
-      wait_until_connection_ready(connection)
+      wait_until_connection_ready
 
       @channel ||= connection.create_channel
       @exchange ||= build_exchange
     end
 
-    def wait_until_connection_ready(conn)
-      Timeout.timeout((conn.heartbeat || 60) * 2) do # 60 seconds is a default Bunny heartbeat
-        loop do
-          return if conn.status == :open && conn.transport.open?
+    def reset_exchange!
+      ensure_connection!
+      @channel = connection.create_channel
+      @exchange = build_exchange
+    end
 
-          sleep 0.001
+    def wait_until_connection_ready
+      Timeout.timeout(recovery_timeout * 2) do
+        loop do
+          return if connection_open? || !connection.automatically_recover?
+
+          sleep 0.01
         end
       end
     rescue Timeout::Error
       # Connection recovery takes too long, let the next interaction fail with error then.
+    end
+
+    def should_start_connection?
+      connection.status == :not_connected || # Lazy connection initialization
+        connection.closed?
+    end
+
+    def connection_can_recover?
+      connection.automatically_recover? && connection.should_retry_recovery?
+    end
+
+    def connection_open?
+      # Do not trust Bunny::Session#open? - it uses :connected & :connecting statuses as "open",
+      # while connection is not actually ready to work.
+      connection.instance_variable_get(:@status_mutex).synchronize do
+        connection.status == :open && connection.transport.open?
+      end
+    end
+
+    def recovery_timeout
+      # 60 seconds is a default heartbeat timeout https://www.rabbitmq.com/heartbeats.html#heartbeats-timeout
+      # Recommended timeout is 5-20 https://www.rabbitmq.com/heartbeats.html#false-positives
+      heartbeat_timeout = [
+        (connection.respond_to?(:heartbeat_timeout) ? connection.heartbeat_timeout : connection.heartbeat) || 60,
+        5
+      ].max
+
+      # Using x2 of heartbeat timeout to get Bunny chance to detect connection failure & try to recover it
+      heartbeat_timeout * 2
     end
 
     def build_connection
@@ -78,6 +124,18 @@ module BunnyPublisher
       return channel.default_exchange if @exchange_name.nil? || @exchange_name == ''
 
       channel.exchange(@exchange_name, @exchange_options)
+    end
+
+    def with_errors_handling
+      yield
+    rescue Bunny::ChannelAlreadyClosed
+      reset_exchange!
+      retry
+    rescue *RETRIABLE_ERRORS => e
+      raise unless connection_can_recover?
+
+      logger.warn { e.inspect }
+      retry
     end
   end
 end